+++ src/app/features/dashboard/all.js
@@ -1,0 +1,14 @@
+define([
+  './dashboardCtrl',
+  './dashboardNavCtrl',
+  './playlistCtrl',
+  './rowCtrl',
+  './sharePanelCtrl',
+  './submenuCtrl',
+  './dashboardSrv',
+  './keybindings',
+  './viewStateSrv',
+  './playlistSrv',
+  './panelSrv',
+  './timeSrv',
+], function () {});