+++ src/app/panels/graph/module.js
@@ -7,9 +7,6 @@ define([
   'moment',
   'components/timeSeries',
   'components/panelmeta',
-  'services/panelSrv',
-  'services/annotationsSrv',
-  'services/datasourceSrv',
   './seriesOverridesCtrl',
   './graph',
   './legend',
@@ -146,10 +143,9 @@ function (angular, app, $, _, kbn, moment, TimeSeries, PanelMeta) {
     };
 
     $scope.dataHandler = function(results) {
-      $scope.panelMeta.loading = false;
-
       // png renderer returns just a url
       if (_.isString(results)) {
+        $scope.panelMeta.loading = false;
         $scope.render(results);
         return;
       }
@@ -164,9 +160,11 @@ function (angular, app, $, _, kbn, moment, TimeSeries, PanelMeta) {
 
       $scope.annotationsPromise
         .then(function(annotations) {
+          $scope.panelMeta.loading = false;
           $scope.seriesList.annotations = annotations;
           $scope.render($scope.seriesList);
         }, function() {
+          $scope.panelMeta.loading = false;
           $scope.render($scope.seriesList);
         });
     };