+++ src/app/routes/all.js
@@ -1,7 +1,0 @@
-define([
-  './pro/dashboard',
-  './pro/admin',
-  './pro/solo-panel',
-  './pro/login',
-],
-function () {});