+++ src/app/routes/backend/all.js
@@ -1,0 +1,6 @@
+define([
+  './dashboard',
+  './admin',
+  './login',
+  './solo-panel',
+], function() {});