+++ src/test/specs/timeSrv-specs.js
@@ -1,8 +1,9 @@
 define([
   'mocks/dashboard-mock',
-  './helpers',
+  'helpers',
   'lodash',
-  'services/timeSrv'
+  'services/timer',
+  'features/dashboard/timeSrv'
 ], function(dashboardMock, helpers, _) {
   'use strict';